requirements.txt
# app
pillow

# test
pytest